# Conflicts:
#	src/main/java/com/javamentor/developer/social/platform/webapp/controllers/v2/GroupControllerV2.java
#	src/main/java/com/javamentor/developer/social/platform/webapp/controllers/v2/UserControllerV2.java

+++ src/main/java/com/javamentor/developer/social/platform/webapp/controllers/v2/PostControllerV2.java
@@ -181,6 +181,12 @@ public class PostControllerV2 {
     public ResponseEntity<?> addLikeToPost(
             @ApiParam(value = "Id поста", example = "1") @PathVariable @NonNull Long postId) {
         User user = userService.getPrincipal();
+
+        Optional<PostLike> optionalPostLike = postLikeService.getPostLikeByPostIdAndUserId(postId, user.getUserId());
+        if (optionalPostLike.isPresent()) {
+            return ResponseEntity.badRequest().body("The Like has already been added");
+        }
+
         Post post = Post.builder().id(postId).build();
         PostLike postLike = new PostLike(user);
         likeService.create(postLike.getLike());
@@ -201,9 +207,11 @@ public class PostControllerV2 {
         User user = userService.getPrincipal();
         Optional<PostLike> optionalLike =
                 postLikeService.getPostLikeByPostIdAndUserId(postId, user.getUserId());
+
         if(!optionalLike.isPresent()) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("The Like has already been deleted");
+            return ResponseEntity.badRequest().body("The Like already been removed");
         }
+
         PostLike postLike = optionalLike.get();
         postLikeService.delete(postLike);
         return ResponseEntity.status(HttpStatus.CREATED)
@@ -218,6 +226,12 @@ public class PostControllerV2 {
     public ResponseEntity<?> addPostToBookmark(
             @ApiParam(value = "Id поста", example = "1") @PathVariable @NonNull Long postId) {
         User user = userService.getPrincipal();
+
+        Optional<Bookmark> optionalBookmark = bookmarkService.getBookmarkByPostIdAndUserId(postId, user.getUserId());
+        if(optionalBookmark.isPresent()) {
+            return ResponseEntity.badRequest().body("The Post has already been added to the bookmark");
+        }
+
         Post post = Post.builder().id(postId).build();
         Bookmark bookmark = Bookmark.builder().user(user).post(post).build();
         bookmarkService.create(bookmark);
@@ -233,6 +247,12 @@ public class PostControllerV2 {
     public ResponseEntity<?> deletePostFromBookmark(
             @ApiParam(value = "Id поста", example = "1") @PathVariable @NonNull Long postId) {
         User user = userService.getPrincipal();
+
+        Optional<Bookmark> optionalBookmark = bookmarkService.getBookmarkByPostIdAndUserId(postId, user.getUserId());
+        if(!optionalBookmark.isPresent()) {
+            return ResponseEntity.badRequest().body("The Post has already been removed from the bookmark");
+        }
+
         bookmarkService.deleteBookmarkByPostIdAndUserId(postId, user.getUserId());
         return ResponseEntity.status(HttpStatus.CREATED)
                 .body(postDtoService.getPostById(postId, user.getUserId()));
